+++ src/app/components/essence-ionic.module.ts
@@ -26,10 +26,5 @@ const components: any[] = [
 	exports: components
 })
 export class EssenceIonicModule {
-	static setAMapKey(config: {apiKey: string, webApiKey: string}) {
-		EssenceIonAMapComponent.webApiKey = config.webApiKey;
-		EssenceIonAMapComponent.apiKey = config.apiKey;
-		return this;
-	};
 }